+++ responsibly/__init__.py
@@ -1,0 +1,15 @@
+# pylint: disable=line-too-long
+
+from responsibly import fairness, we
+
+
+__project__ = 'responsibly'
+__description__ = 'Toolkit for Auditing and Mitigating Bias and Fairness of Machine Learning Systems 🔎🤖🧰'
+__url__ = 'http://docs.responsibly.ai'
+__version__ = '0.1.0'
+__author__ = 'Shlomi Hod'
+__author_email__ = 'shlomi.hod@gmail.com'
+__license__ = 'MIT'
+__copyright__ = 'Copyright 2019 Shlomi Hod'
+
+VERSION = "{0} v{1}".format(__project__, __version__)